+++ src/webpack/model/interface/id.ts
@@ -1,0 +1,5 @@
+export interface id {
+    server: string;
+    user: string;
+    _serialized: string;
+  }