+++ src/webpack/model/interface/index.ts
@@ -3,3 +3,6 @@ export { sendOptions } from './send-layer';
 export { interfaceChange } from './interface-change';
 export { InterfaceQrcode } from './qrcode';
 export { InterfaceScope } from './scope';
+export { InterfaceHost } from './host';
+export { id } from './id';
+export { contact } from './contact';